+++ database/tx.go
@@ -14,14 +14,14 @@ func NewAccount(value string) Account {
 
 type Tx struct {
 	From  		Account 	`json:"from"`
-	Value 		uint    	`json:"value"`
+	Value 		uint64    	`json:"value"`
 	Repository  string  	`json:"repository"`
 	Commit 		string 		`json:"commit"`
 	prevCommit 	string 		`json:"prevCommit"`
 	Time  		uint64  	`json:"time"`
 }
 
-func NewTx(from Account, value uint, repository string, commit string, prevCommit string) Tx {
+func NewTx(from Account, value uint64, repository string, commit string, prevCommit string) Tx {
 	return Tx{from, value, repository, commit, prevCommit, uint64(time.Now().Unix())}
 }
